requirements.txt
python-jobspy
openai
pandas
python-docx
openpyxl